+++ lib/active_record/mysql/enum/active_record_helper.rb
