+++ lib/active_record/mysql/enum/enum_adapter.rb
