+++ lib/active_record/mysql/enum/enum_column_adapter.rb
