+++ lib/active_record/mysql/enum/enum_type.rb
