+++ lib/active_record/mysql/enum/quoting.rb
