+++ lib/active_record/mysql/enum/schema_definitions.rb
