+++ lib/active_record/mysql/enum/validations.rb
