--- lib/enum_column3.rb
@@ -1,1 +1,0 @@
-require 'enum_column'
